Move services to separate files.
