+++ custom_components/rohlikcz/__init__.py
@@ -2,17 +2,14 @@
 from __future__ import annotations
 
 import logging
-import voluptuous as vol
-from typing import Any, Dict, List
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_EMAIL, CONF_PASSWORD
-from homeassistant.core import HomeAssistant, ServiceCall
-import homeassistant.helpers.config_validation as cv
-from homeassistant.exceptions import HomeAssistantError
+from homeassistant.core import HomeAssistant
 
 from .const import DOMAIN
 from .hub import RohlikAccount
+from .services import register_services
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -56,145 +53,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     return unload_ok
 
 
-def register_services(hass: HomeAssistant) -> None:
-    """Register services for the Rohlik integration."""
-
-    async def async_add_to_cart_service(call: ServiceCall) -> List[int]:
-        """Add product to cart service."""
-        config_entry_id = call.data[ATTR_CONFIG_ENTRY_ID]
-        product_id = call.data[ATTR_PRODUCT_ID]
-        quantity = call.data[ATTR_QUANTITY]
-
-        if config_entry_id not in hass.data[DOMAIN]:
-            raise HomeAssistantError(f"Config entry {config_entry_id} not found")
-
-        account = hass.data[DOMAIN][config_entry_id]
-        try:
-            result = await account.add_to_cart(product_id, quantity)
-            _LOGGER.info(f"Product added to cart for {account.name}: {result}")
-            return result
-        except Exception as err:
-            _LOGGER.error(f"Failed to add product to cart: {err}")
-            raise HomeAssistantError(f"Failed to add product to cart: {err}")
-
-    async def async_search_product_service(call: ServiceCall) -> Dict[str, Any]:
-        """Search for a product and return results."""
-        config_entry_id = call.data[ATTR_CONFIG_ENTRY_ID]
-        product_name = call.data[ATTR_PRODUCT_NAME]
-
-        if config_entry_id not in hass.data[DOMAIN]:
-            raise HomeAssistantError(f"Config entry {config_entry_id} not found")
-
-        account = hass.data[DOMAIN][config_entry_id]
-        try:
-            result = await account.search_product(product_name)
-            return result or {}
-        except Exception as err:
-            _LOGGER.error(f"Failed to search for product: {err}")
-            raise HomeAssistantError(f"Failed to search for product: {err}")
-
-    async def async_search_and_add_product_service(call: ServiceCall) -> Dict[str, Any]:
-        """Search for a product and return results."""
-        config_entry_id = call.data[ATTR_CONFIG_ENTRY_ID]
-        product_name = call.data[ATTR_PRODUCT_NAME]
-        quantity = call.data[ATTR_QUANTITY]
-
-        if config_entry_id not in hass.data[DOMAIN]:
-            raise HomeAssistantError(f"Config entry {config_entry_id} not found")
-
-        account = hass.data[DOMAIN][config_entry_id]
-        try:
-            result = await account.search_and_add(product_name,quantity)
-            return result or {}
-        except Exception as err:
-            _LOGGER.error(f"Failed to search for product: {err}")
-            raise HomeAssistantError(f"Failed to search for product: {err}")
-
-
-    async def async_get_shopping_list_service(call: ServiceCall) -> Dict[str, Any]:
-        """Get shopping list by ID."""
-        config_entry_id = call.data[ATTR_CONFIG_ENTRY_ID]
-        shopping_list_id = call.data[ATTR_SHOPPING_LIST_ID]
-
-        if config_entry_id not in hass.data[DOMAIN]:
-            raise HomeAssistantError(f"Config entry {config_entry_id} not found")
-
-        account = hass.data[DOMAIN][config_entry_id]
-        try:
-            result = await account.get_shopping_list(shopping_list_id)
-            return result
-        except Exception as err:
-            _LOGGER.error(f"Failed to get shopping list: {err}")
-            raise HomeAssistantError(f"Failed to get shopping list: {err}")
-
-    async def async_get_cart_service(call: ServiceCall) -> Dict[str, Any]:
-        """Get shopping cart content."""
-        config_entry_id = call.data[ATTR_CONFIG_ENTRY_ID]
-
-        if config_entry_id not in hass.data[DOMAIN]:
-            raise HomeAssistantError(f"Config entry {config_entry_id} not found")
-
-        account = hass.data[DOMAIN][config_entry_id]
-        try:
-            result = await account.get_cart_content()
-            return result
-        except Exception as err:
-            _LOGGER.error(f"Failed to get cart content: {err}")
-            raise HomeAssistantError(f"Failed to get get cart content: {err}")
-
-    # Register the services
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_ADD_TO_CART,
-        async_add_to_cart_service,
-        schema=vol.Schema({
-            vol.Required(ATTR_CONFIG_ENTRY_ID): cv.string,
-            vol.Required(ATTR_PRODUCT_ID): cv.positive_int,
-            vol.Required(ATTR_QUANTITY, default=1): cv.positive_int,
-        }),
-        supports_response=True
-    )
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_SEARCH_PRODUCT,
-        async_search_product_service,
-        schema=vol.Schema({
-            vol.Required(ATTR_CONFIG_ENTRY_ID): cv.string,
-            vol.Required(ATTR_PRODUCT_NAME): cv.string,
-        }),
-        supports_response=True
-    )
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_SEARCH_AND_ADD_PRODUCT,
-        async_search_and_add_product_service,
-        schema=vol.Schema({
-            vol.Required(ATTR_CONFIG_ENTRY_ID): cv.string,
-            vol.Required(ATTR_PRODUCT_NAME): cv.string,
-            vol.Required(ATTR_QUANTITY): cv.positive_int,
-        }),
-        supports_response=True
-    )
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_GET_SHOPPING_LIST,
-        async_get_shopping_list_service,
-        schema=vol.Schema({
-            vol.Required(ATTR_CONFIG_ENTRY_ID): cv.string,
-            vol.Required(ATTR_SHOPPING_LIST_ID): cv.string,
-        }),
-        supports_response=True
-    )
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_GET_CART_CONTENT,
-        async_get_cart_service,
-        schema=vol.Schema({
-            vol.Required(ATTR_CONFIG_ENTRY_ID): cv.string,
-        }),
-        supports_response=True
-    )
